Rebase

+++ src/components/BrandDetails/Product.js
@@ -245,7 +245,7 @@ function Product() {
         Object.values(begValue.orderList).map((product) => {
           let productPriceStr = product.product.salesPrice;
           let productQuantity = product.quantity;
-          let productPrice = parseInt(productPriceStr || 0);
+          let productPrice = parseFloat(productPriceStr || 0);
           bagPrice += productPrice * productQuantity;
         });
         setAlert(0);
